requirements.txt
tqdm
Pillow
PyPDF2